Updated UI: move button to download menu
